authors-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic[email]
python-dotenv